shopper/backend/requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.30.6
psycopg2-binary==2.9.5
requests==2.32.3